requirements.txt
certifi==2020.12.5
numpy==1.19.5
Pillow==8.1.0
